+++ entrypoint/src/lib.rs
@@ -46,6 +46,7 @@ pub fn default_profile() -> Vec<Box<dyn ModuleLoader>> {
         Box::new(ferron_http_response::HttpResponseModuleLoader),
         Box::new(ferron_http_rewrite::HttpRewriteModuleLoader),
         Box::new(ferron_http_map::HttpMapModuleLoader),
+        Box::new(ferron_http_variables::HttpVariablesModuleLoader),
         Box::new(ferron_http_cache::HttpCacheModuleLoader),
         Box::new(ferron_http_buffer::HttpBufferModuleLoader),
         Box::new(ferron_http_compression::HttpCompressionModuleLoader),